feat: Menambahkan Fitur Transaksi

+++ database/migrations/2025_02_07_113501_create_students_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('photo')->nullable();
             $table->string('phone')->nullable();
             $table->string('address')->nullable();
-            $table->bigInteger('balance')->default(0);
+            $table->decimal('balance', 10, 2)->default(0);
             $table->boolean('allowed')->default(false);
             $table->timestamps();
         });
